src/agent/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Acuvity AI Chatbot"
authors = [
    {name = "Satyam Sinha",email = "satyam@acuvity.ai"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "alembic (>=1.12.0)",
    "bcrypt (==3.2.2)",
    "descope (>=0.1.0)",
    "fast-agent-mcp @ git+https://github.com/acuvity/fast-agent.git",
    "fastapi (>=0.103.1)",
    "httpx (>=0.25.0)",
    "opentelemetry-api (>=1.31.1,<2.0.0)",
    "openinference-instrumentation-mcp>=1.1.0",
    "opentelemetry-instrumentation-fastapi (>=0.52b1,<0.53)",
    "opentelemetry-instrumentation-httpx (>=0.52b1,<0.53)",
    "opentelemetry-instrumentation-threading>=0.52b1",
    "passlib (>=1.7.4)",
    "psycopg2-binary (>=2.9.6)",
    "pydantic (>=2.4.2)",
    "pydantic-settings (>=2.0.3)",
    "pytest (>=7.4.2)",
    "python-dotenv (>=1.0.0)",
    "python-jose[cryptography] (>=3.3.0)",
    "python-multipart (>=0.0.6)",
    "sqlalchemy (>=2.0.21)",
    "uvicorn (>=0.23.2)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
